Fetch set data in server component before render of client
