+++ src/app/sets/[id]/page.tsx
@@ -1,147 +1,18 @@
-"use client";import { useEffect, useState } from "react";
-import Cards from "./Cards";
-import EditCardModal from "./EditCardModal";
-import { useUser } from "@/app/context/UserProvider";
-import { toast } from "react-toastify";
-import useFavButton from "@/app/hooks/useFavButton";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faHeart as fillHeart } from "@fortawesome/free-solid-svg-icons";
-import { faHeart as emptyHeart } from "@fortawesome/free-regular-svg-icons";
-import "./ViewSet.scss";
 import { getSet } from "@/app/lib/api";
-import { Card, FullSet } from "@/app/lib/definitions";
-import Loading from "@/app/loading";
-import Link from "next/link";
-import { useRouter } from "next/navigation";
+import { AxiosError, isAxiosError } from "axios";
+import ViewSet from "./ViewSet";
+import { notFound } from "next/navigation";
 
-export default function Page({ params }: { params: { id: string } }) {
-  const router = useRouter();
-  const { user, favoriteSets } = useUser();
-  const { isLiked, checkLiked, toggleLike } = useFavButton();
+export default async function Page({ params }: { params: { id: string } }) {
+  try {
+    const setData = await getSet(params.id);
 
-  const [setData, setSetData] = useState<FullSet | null>(null);
-  const [isLoading, setIsLoading] = useState(true);
-  const [showEditModal, setShowEditModal] = useState(false);
-  const [editingCard, setEditingCard] = useState<Card | null>(null);
-
-  useEffect(() => {
-    getSet(params.id)
-      .then(setSetData)
-      .catch((err) => {
-        toast.error(err);
-      })
-      .finally(() => setIsLoading(false));
-    // check whether the current set is liked by the logged in user
-    checkLiked(favoriteSets, Number(params.id));
-  }, [params.id]);
-
-  const handleCardEdit = (card: Card) => {
-    setEditingCard(card);
-    setShowEditModal(true);
-  };
-
-  const closeEditModal = () => {
-    setEditingCard(null);
-    setShowEditModal(false);
-  };
-
-  const handleCardUpdate = (updatedCard: Card) => {
-    setSetData((prevData: any) => {
-      const newCards =
-        prevData &&
-        prevData.set &&
-        prevData.cards.map((card: Card) =>
-          card.id === updatedCard.id ? updatedCard : card
-        );
-      return { ...prevData, cards: newCards };
-    });
-    setShowEditModal(false);
-  };
-
-  if (isLoading) {
-    return <Loading />;
+    return <ViewSet fullSetData={setData} />;
+  } catch (err: any | AxiosError) {
+    if (isAxiosError(err) && err.response?.status === 404) {
+      notFound();
+    } else {
+      console.log(err);
+    }
   }
-
-  if (!setData) {
-    return (
-      <main className="mt-8 text-center">
-        <h2>Set Not Found</h2>
-      </main>
-    );
-  }
-
-  const { set, cards } = setData;
-
-  if (set.private && (!user || user.id !== set.user_id)) {
-    return (
-      <main className="mt-8 text-center">
-        <h2 className="text-xl mb-4">This set is marked as private.</h2>
-        <button className="btn" onClick={() => router.back()}>
-          Return
-        </button>
-      </main>
-    );
-  }
-
-  return (
-    <main className="py-5 px-4 md:p-0 max-w-4xl mx-auto ">
-      <section className="flex justify-between items-center md:items-end gap-2">
-        <div className="md:flex items-end gap-2">
-          <h1 className="text-[2rem] font-bold mb-2 md:mb-0 md:text-4xl">
-            {set.title}
-          </h1>
-          <h2 className="bg-color-3 rounded-md p-2 text-base inline-block mb-2 md:mb-0">
-            {set.category_name}
-          </h2>
-          {user && (
-            <button
-              className="text-3xl inline-block align-middle ml-2"
-              onClick={() => toggleLike(set)}
-            >
-              {isLiked ? (
-                <FontAwesomeIcon
-                  icon={fillHeart}
-                  className="icon-primary heart-icon"
-                />
-              ) : (
-                <FontAwesomeIcon
-                  icon={emptyHeart}
-                  className="icon-primary heart-icon"
-                />
-              )}
-            </button>
-          )}
-        </div>
-        {user && user.id === set.user_id && (
-          <Link className="btn" href={`/sets/edit/${params.id}`}>
-            Edit Set
-          </Link>
-        )}
-      </section>
-
-      <Cards
-        cards={cards}
-        isSetOwner={user && user.id === set.user_id}
-        onEdit={handleCardEdit}
-      />
-
-      {/* Edit Card Modal */}
-      {editingCard && (
-        <EditCardModal
-          show={showEditModal}
-          onHide={closeEditModal}
-          card={editingCard}
-          onUpdate={handleCardUpdate}
-        />
-      )}
-
-      <section className="px-0 flex gap-2 md:px-8 justify-between">
-        <p className="p-4 text-lg font-bold basis-3/12">{set.username}</p>
-        <div className="p-4 basis-9/12">
-          <h3 className="text-lg">Description:</h3>
-          <p>{set.description}</p>
-        </div>
-      </section>
-    </main>
-  );
 }
